split hit.c file
